Move font functionality to org.jetbrains.skiko.awt.font
