+++ skiko/src/awtMain/kotlin/org/jetbrains/skiko/JvmUtils.kt
@@ -1,4 +1,4 @@
-package org.jetbrains.skiko.context
+package org.jetbrains.skiko
 
 internal fun isRunningOnJetBrainsRuntime() =
     System.getProperty("java.vendor")
