+++ skiko/src/awtMain/kotlin/org/jetbrains/skiko/awt/font/FontFamily.kt
@@ -1,4 +1,4 @@
-package org.jetbrains.skiko
+package org.jetbrains.skiko.awt.font
 
 import org.jetbrains.skia.*
