+++ skiko/src/awtTest/kotlin/org/jetbrains/skiko/awt/font/AwtFontManagerTest.kt
@@ -1,4 +1,4 @@
-package org.jetbrains.skiko
+package org.jetbrains.skiko.awt.font
 
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
@@ -6,15 +6,13 @@ import org.jetbrains.skia.Data
 import org.jetbrains.skia.FontStyle
 import org.jetbrains.skia.Typeface
 import org.jetbrains.skia.tests.makeFromResource
-import org.jetbrains.skiko.AwtFontUtils.fontFamilyName
-import org.jetbrains.skiko.AwtFontUtils.resolvePhysicalFontNameOrNull
-import org.jetbrains.skiko.context.isRunningOnJetBrainsRuntime
+import org.jetbrains.skiko.*
+import org.jetbrains.skiko.awt.font.AwtFontUtils.fontFamilyName
+import org.jetbrains.skiko.isRunningOnJetBrainsRuntime
 import org.jetbrains.skiko.tests.runTest
 import org.junit.Assume
-import org.junit.Ignore
 import org.junit.Test
 import java.awt.GraphicsEnvironment
-import kotlin.io.path.createTempFile
 import kotlin.io.path.writeBytes
 import kotlin.test.*
 
@@ -83,7 +81,7 @@ class AwtFontManagerTest {
         val skiaFonts = awtFonts.map { it.toSkikoTypefaceOrNull(fontManager) }
 
         fun String.resolveFontFamily() = if (FontFamilyKey(this) in FontFamilyKey.Awt.awtLogicalFonts) {
-            resolvePhysicalFontNameOrNull(this) ?: this
+            AwtFontUtils.resolvePhysicalFontNameOrNull(this) ?: this
         } else {
             this
         }
@@ -174,7 +172,7 @@ class AwtFontManagerTest {
             "The font we're trying to add must not already be loaded"
         )
 
-        val fontFile = createTempFile("awtfontmanagertest", "testfont")
+        val fontFile = kotlin.io.path.createTempFile("awtfontmanagertest", "testfont")
         withContext(Dispatchers.IO) {
             val fontBytes = Thread.currentThread()
                 .contextClassLoader
@@ -241,7 +239,7 @@ class AwtFontManagerTest {
         fontManager.addCustomFontTypeface(Typeface.makeFromResource("./fonts/JetBrainsMono-Regular.ttf"))
         fontManager.addCustomFontResource("./fonts/JetBrainsMono-Italic.ttf")
 
-        val fontFile = createTempFile("awtfontmanagertest", "testfont")
+        val fontFile = kotlin.io.path.createTempFile("awtfontmanagertest", "testfont")
         withContext(Dispatchers.IO) {
             val fontBytes = Thread.currentThread()
                 .contextClassLoader
